+++ slash/core/runnable_test_factory.py
@@ -50,9 +50,9 @@ class RunnableTestFactory(object):
 
         Do not override this method directly. Use :func:`.RunnableTestFactory._generate_tests` instead.
         """
-        for address_in_factory, test in self._generate_tests(fixture_store):
+        for address_in_factory, variation, test in self._generate_tests(fixture_store):
             assert test.__slash__ is None
-            test.__slash__ = Metadata(self, test, address_in_factory)
+            test.__slash__ = Metadata(self, test, address_in_factory, variation=variation.representation)
             yield test
 
     def _generate_tests(self, fixture_store):