+++ tests/test_fixture_string_representation.py
@@ -17,9 +17,9 @@ def test_representation_parameters(fixture_store, parametrized_func, params, par
     cartesian_product = list(iter_cartesian_dicts(params))
 
     assert len(variations) == len(cartesian_product)
-    variation_names = set(v.representation for v in variations)
+    variation_names = set(str(v.representation) for v in variations)
     assert len(variation_names) == len(cartesian_product)
-    assert variation_names == set(', '.join('{0}={1}'.format(name, combination[name]) for name in param_names)
+    assert variation_names == set(', '.join('{0}={1}'.format(name, combination[name]) for name in sorted(param_names))
                                   for combination in cartesian_product)
 
 def test_representation_fixtures(fixture_store):
@@ -49,7 +49,7 @@ def test_representation_fixtures(fixture_store):
 
     assert len(variations) == len(first_param_values) * len(second_param_values)
 
-    assert set(variation.representation for variation in variations) == set(
+    assert set(str(variation.representation) for variation in variations) == set(
         'first_fixture=first_fixture{0}, second_fixture=second_fixture{1}'.format(i, j)
         for i, j in itertools.product(range(len(first_param_values)), range(len(second_param_values))))